autoML/autoML-cloud_fxn/cloud_fxn/requirements.txt
yfinance==0.1.70
google-cloud-storage==2.1.0 
google-cloud-bigquery==2.31.0
google-cloud-aiplatform==1.13.1